+++ services/healthcareapis/mgmt/2020-03-15/healthcareapis/CHANGELOG.md
@@ -1,2 +1,10 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. Operation.IsDataAction
+1. Operation.Properties